Solarica/hint_from_chatgpt/pvpm-platform/apps/local-reader/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-dotenv==1.0.1
pyserial==3.5
requests==2.32.3
